requirements.txt
tensorflow==1.14
imutils
numpy
pillow
lxml
jupyter
matplotlib
opencv-python
cython
